add score chart link to form
