+++ app/_components/ui/(form)/MainForm.tsx
@@ -9,6 +9,8 @@ import { FormProvider, useForm } from "react-hook-form";
 import ExerciseFields from "./(controls)/ExerciseFields";
 import GenderRadio from "./(controls)/GenderRadioButtons";
 import Score from "./(score)/Score";
+import { formatAgeGroup } from "@/app/_util/helpers";
+import Link from "next/link";
 
 export type formType = {
   gender: string;
@@ -37,7 +39,10 @@ const MainForm = () => {
     },
   });
   const { setScores } = useScoreContext();
+  const gender = methods.watch("gender");
+  const ageGroup = methods.watch("ageGroup");
 
+  const formattedAgeGroup = formatAgeGroup(ageGroup);
   async function onSubmit(data: formType) {
     const res = await calculateTotalScoresWithMinimumCheck(data);
     setScores(res);
@@ -51,6 +56,18 @@ const MainForm = () => {
       >
         <GenderRadio />
         <AgeGroupSelect />
+        {gender && ageGroup && (
+          <p className="text-center text-lg">
+            View the {gender} / {ageGroup} score chart{" "}
+            <Link
+              href={`https://hnnyotwjhikrytqynjyk.supabase.co/storage/v1/object/public/usafptcalculator/scorechart_${gender}_${formattedAgeGroup}.pdf`}
+              target="_blank"
+              className="text-primary hover:text-primary/50"
+            >
+              here
+            </Link>
+          </p>
+        )}
         <ExerciseFields />
         <FormButtons />
       </form>
